DialogueContainer の CanvasGroup 透明度処理を専用クラスに移す
