+++ Assets/_MAIN/Scripts/Core/CanvasGroupController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7c5d19a973aa54549bcfb4687f8cbaea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
