+++ Assets/_MAIN/Scripts/Core/DisplayDialogue/TMProDisplayContainer/DialogueContainer.cs
@@ -14,8 +14,6 @@ namespace Core.DisplayDialogue
     [System.Serializable]
     public class DialogueContainer
     {
-        private const float DEFAULT_FADE_SPEED = 1f;
-
         /// <summary>
         /// Layers.4 - Dialogue
         /// ダイアログレイヤ全体（名前・ダイアログを含む）表示・非表示などに使える
@@ -33,15 +31,12 @@ namespace Core.DisplayDialogue
         /// </summary>
         public TextMeshProUGUI dialogueText;
 
-        private CanvasGroup canvasGroup => rootGameObject.GetComponent<CanvasGroup>();
+        private CanvasGroupController canvasGroupController;
 
-        private Coroutine showingCoroutine;
-        private Coroutine hidingCoroutine;
-
-        private bool isShowing => showingCoroutine != null;
-        private bool isHiding => hidingCoroutine != null;
-        private bool isFading => isShowing || isHiding;
-        public bool isVisible => isShowing || canvasGroup.alpha > 0f;
+        public void Initialize(MonoBehaviour owner)
+        {
+            canvasGroupController = new CanvasGroupController(owner, rootGameObject.GetComponent<CanvasGroup>());
+        }
 
         public void ApplyCharacterConfig(CharacterConfig characterConfig,
             DialogueSystemConfigurationSO dialogueSystemConfig)
@@ -52,41 +47,8 @@ namespace Core.DisplayDialogue
             dialogueText.fontSize = characterConfig.dialogueFontSize * dialogueSystemConfig.dialogueFontScale;
         }
 
-        public Coroutine Show()
-        {
-            if (isShowing) return null;
-            if (isHiding)
-            {
-                DialogueSystemController.instance.StopCoroutine(hidingCoroutine);
-                hidingCoroutine = null;
-            }
-
-            return showingCoroutine = DialogueSystemController.instance.StartCoroutine(Fading(1f));
-        }
-
-        public Coroutine Hide()
-        {
-            if (isHiding) return null;
-            if (isShowing)
-            {
-                DialogueSystemController.instance.StopCoroutine(showingCoroutine);
-                showingCoroutine = null;
-            }
-
-            return hidingCoroutine = DialogueSystemController.instance.StartCoroutine(Fading(0f));
-        }
-
-        public IEnumerator Fading(float targetAlpha)
-        {
-            var cg = canvasGroup;
-
-            while (!Mathf.Approximately(cg.alpha, targetAlpha))
-            {
-                cg.alpha = Mathf.MoveTowards(cg.alpha, targetAlpha, Time.deltaTime * DEFAULT_FADE_SPEED);
-                yield return null;
-            }
-
-            showingCoroutine = hidingCoroutine = null;
-        }
+        public bool isVisible => canvasGroupController.isVisible;
+        public Coroutine Show() => canvasGroupController.Show();
+        public Coroutine Hide() => canvasGroupController.Hide();
     }
 }
